refactor: Add auth module
Add dedicated module to enhance cohesion and separation of concerns .
Migrate all auth-related code to auth module to increase modularity.

Chores:
  - Modify import dependencies in required files.
  - Create a dedicated file to store constants.
  - Update tests to adapt to code refactor.
